Does not lose split alignments.  Now to work on reset.
